% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recurrence_bin_functions.R
\name{bin}
\alias{bin}
\title{Bin Recurrence Times}
\usage{
bin(recurrences, binwidth = 0.001)
}
\arguments{
\item{recurrences}{Numeric vector representing recurrence times.}

\item{binwidth}{Numeric. Default = 0.001 (0.001 second or 1 ms). This must
have the same significant digits as the recurrences parameter.}
}
\value{
data frame containing frequency data.
}
\description{
Discretizes recurrence times vector into a data frame according to specified
binwidth parameter. A recurrence time of "0" indicates two events happening
simultaneously. The "0" bin is treated as the center.
}
\keyword{action}
\keyword{bin,}
\keyword{discharge}
\keyword{group,}
\keyword{motor}
\keyword{potentials,}
\keyword{trains}
\keyword{unit,}
