% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/motor_unit_data.R
\docType{data}
\name{motor_unit_data}
\alias{motor_unit_data}
\title{Motor unit time series}
\format{Data frame with 30,000 rows and 3 columns
\describe{
\item{Time}{time in seconds}
\item{motor_unit_1}{Binary indicator of discharge (1) or not (0)}
\item{motor_unit_2}{Binary indicator of discharge (1) or not (0)}
}}
\usage{
data(motor_unit_data)
}
\description{
Example time series of 2 motor units' discharges for the calculation of
synchronization.
}
\keyword{datasets}
