% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mu_synch_function.R
\name{mu_synch}
\alias{mu_synch}
\title{Determination of Motor Unit Synchronization by Various Methods of Cross
Correlation Histogram Examination}
\usage{
mu_synch(motor_unit_1, motor_unit_2, method = "Visual", order = 1,
  binwidth = 0.001, plot = F)
}
\arguments{
\item{motor_unit_1}{Numeric vectors of strictly increasing numbers denoting
sequential discharge times of a motor unit or neuron or any strictly
increasing point process.}

\item{motor_unit_2}{Numeric vectors of strictly increasing numbers denoting
sequential discharge times of a motor unit or neuron or any strictly
increasing point process.}

\item{method}{Character vector indicating which methods of peak detection to
use when quantifying synchronization. "Visual", "Zscore", and "Cumsum" are
the options. Default is Visual.}

\item{order}{Numeric as a positive integer for the number of forward and
backward orders for calculating recurrence times. Default = 1.}

\item{binwidth}{Numeric as a positive for the bin allocation size for
histogram computation. Default = 0.001 or 1 ms.}

\item{plot}{T/F logical for outputting the cross correlation histogram.
Default is FALSE.}
}
\value{
A list of lists containing motor unit data (the names of each
  discharge train used, number of discharges, the interspike intervals (ISI),
  mean ISI, and the recurrence times associated with each order) and
  synchronization indices associated with chosen methods.
  CIS = frequency of synchronized discharges.
  k' = ratio of total discharges in peak to expected discharges in peak.
  k'-1 = ratio of synchronized discharges to expected discharges in peak.
  S = ratio of synchronized discharges to total number of discharges of both
  motor units.
  E = ratio of synchronized discharges to non-synchronized discharges.
  SI = ratio of synchronized discharges to reference motor unit discharges.
}
\description{
Calculates the time-domain synchronization indices CIS, k',
  k'-1, S, E, SI (detailed below) between the two input motor unit discharge
  trains. One or more methods of peak determination can be chosen ("Visual",
  "Zscore", or "Cumsum"). Chosen method functions are called individually and
  detailed descriptions are documented in their respective help files. Motor
  unit characteristic data and interspike intervals are automatically output
  with mu_synch.
}
\references{
Keen, D.A., Chou, L., Nordstrom, M.A., Fuglevand, A.J. (2012)
  Short-term Synchrony in Diverse Motor Nuclei Presumed to Receive Different
  Extents of Direct Cortical Input. Journal of Neurophysiology 108: 3264-3275

  Nordstrom, M.A., Fuglevand, A.J., Enoka, R.M. (1992) Estimating the
  Strength of Common Input to Human Motoneurons from the Cross-Correlogram.
  Journal of Physiology 453, pp. 547-574

  DeFreitas, J.M., Beck, T.W., Xin, Y., Stock, M.S. (2013) Synchronization of
  Low- and High-Threshold Motor Units. Muscle & Nerve DOI 10.1002/mus.23978
}
\seealso{
visual_mu_synch, zscore_mu_synch, cumsum_mu_synch
}
\keyword{Zscore,}
\keyword{cumulative}
\keyword{motor}
\keyword{recurrence,}
\keyword{sigmax,}
\keyword{sum}
\keyword{synchronization,}
\keyword{unit,}
\keyword{visual,}
