% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recurrence_bin_functions.R
\name{plot_bins}
\alias{plot_bins}
\title{Plot binned recurrence times as a histogram}
\usage{
plot_bins(binned_data)
}
\arguments{
\item{binned_data}{data frame containing frequency data produced by bin().}
}
\description{
Takes data frame containing bins labels and frequency data to produce
histogram.
}
\keyword{bin,}
\keyword{discharge}
\keyword{histogram}
\keyword{motor}
\keyword{plot,}
\keyword{trains,}
\keyword{unit,}
