% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recurrence_bin_functions.R
\name{recurrence_intervals}
\alias{recurrence_intervals}
\title{Calculate Recurrence Intervals Between Motor Units or Neuron Discharge Trains}
\usage{
recurrence_intervals(motor_unit_1, motor_unit_2, order = 1)
}
\arguments{
\item{motor_unit_1, motor_unit_2}{Numeric vectors of strictly increasing
numbers denoting sequential discharge times of a motor unit or neuron or
any strictly increasing point process.}

\item{order}{Numeric as a positive integer for the number of forward and
backward orders for calculating recurrence times. Default = 1.}
}
\value{
A list of lists containing the names of each discharge train used,
  number of discharges, the interspike intervals (ISI), mean ISI, and the
  recurrence times associated with each order.
}
\description{
Take two vectors representing time points of motor unit or neuron discharges
and calculate multi-order recurrence times between them. This function will
return recurrence times in whatever unit is input but will only accept
numeric vectors (e.g. 0.01 sec, 25 ms, or 17.5 minutes).
}
\keyword{action}
\keyword{discharge}
\keyword{motor}
\keyword{potentials,}
\keyword{recurrence,}
\keyword{trains}
\keyword{unit,}
